cloud_storage: make topic manifest state library
The underlying serializable state for topic manifests (for topics from
versions of Redpanda that serialize topic manifests with serde) can be
pulled out and reused by others who want to upload topic manifests (e.g.
cloud topics), without having to depend on all the base_manifest and
legacy parsing code in topic_manifest.cc. This commit pulls this out
into its own tiny header.
